Add external mismatch parser that can be tested separately
